requirements.txt
sister
matplotlib
tensorflow-gpu
tensorflow_hub
